Remove mogenius from deployment guide
Hi there, mogenius doesn't offer free cloud resources anymore. It is still possible to deploy containers on the free plan but it requires setting up a local Kubernetes. I suppose this is not a use case for the Whatsapp bot so I suggest to remove mogenius from the deployment guide.